deps: Switch to GitHub version of react-quill
The onChange handler exposed by react-quill sends values in HTML format
by default, which is not desired. The 4th argument to this callback
actually exposes a read-only version of the editor, so getText() can be
used to query the text value. However, due to a bug, the version of the
library in npm omits the `return` keyword, resulting in the function
always returning `undefined`. Using GitHub master until a new version is
published.

See https://github.com/zenoamaro/react-quill/issues/88